@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
+    "@angular/animations": "^5.1.0",
     "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
+    "@angular/common": "^5.1.0",
+    "@angular/compiler": "^5.1.0",
+    "@angular/core": "^5.1.0",
+    "@angular/forms": "^5.1.0",
+    "@angular/http": "^5.1.0",
     "@angular/material": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/platform-browser": "^5.1.0",
+    "@angular/platform-browser-dynamic": "^5.1.0",
+    "@angular/router": "^5.1.0",
     "@ngrx/core": "^1.2.0",
     "@ngrx/effects": "^4.1.0",
     "@ngrx/router-store": "^4.1.0",
@@ -30,20 +30,19 @@
     "@ngrx/store-devtools": "^4.1.1",
     "@swimlane/ngx-charts": "^6.1.0",
     "@swimlane/ngx-dnd": "^3.1.0",
+    "angular-file-dropzone": "^2.1.1",
     "core-js": "^2.4.1",
     "d3": "^4.11.0",
-    "ng2-file-upload": "^1.3.0",
     "ngrx-store-freeze": "^0.2.0",
     "ngrx-store-logger": "^0.2.0",
-    "ngx-file-drop": "^2.0.1",
     "reselect": "^3.0.1",
-    "rxjs": "^5.5.2",
+    "rxjs": "^5.5.5",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular/cli": "^1.6.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular/compiler-cli": "^5.1.0",
+    "@angular/language-service": "^5.1.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -59,6 +58,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "typescript": "^2.6.2"
   }
 }
