@@ -11,7 +11,7 @@ export class HttpApiRedirectorService implements HttpInterceptor {
 
 	intercept(req: HttpRequest<any>, next: HttpHandler): Observable<any> {
 		const newReq = req.clone({ url: `${this.baseUrl}/${req.url}`});
-		return next.handle(newReq);
+		return next.handle(req);
 	}
 
 }
