@@ -1,5 +1,4 @@
 import { Component, OnInit, EventEmitter, Output, Input } from '@angular/core';
-import { UploadEvent,  } from 'ngx-file-drop';
 
 @Component({
 	selector: 'img-input-app',
@@ -12,25 +11,15 @@ export class ImgInputComponent implements OnInit {
 	style = 'img-file-drop';
 	fileHover: boolean;
 
+
 	constructor() { }
 
 	ngOnInit() {
 	}
 
-	onFileOver() {
+	onDragOver(event) {
 		this.fileHover = true;
-	}
-
-	onFileDrop(event: UploadEvent) {
-		const files = event.files.map(f => f.fileEntry);
-		this.imgsAdded.emit(files);
-		// const reader = new FileReader();
-
-		// reader.onloadend = function (e) {
-		// 	// placeholder until image is ready
-		// 	self.imgAdded.emit( { id: fileItem.id, data: reader.result, pending: true });
-		// };
-		// reader.readAsDataURL(fileItem._file);
+		event.preventDefault();
 	}
 
 	onFileLeave() {
