@@ -3,17 +3,17 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ImgInputComponent } from './components/img-input/img-input.component';
 import { FileInputComponent } from './components/file-input/file-input.component';
-import { FileDropModule } from 'ngx-file-drop';
 import { FileUploader2 } from './services/file-uploader2.service';
+import { FileDropDirective } from './directives/file-drop.directive';
 
 @NgModule({
 	imports: [
 		CommonModule,
-		FileDropModule
 	],
 	declarations: [
 		ImgInputComponent,
-		FileInputComponent
+		FileInputComponent,
+		FileDropDirective
 	],
 	providers: [ FileUploader2 ],
 	exports: [
