@@ -3,6 +3,15 @@ import { AppErrorsEffects } from './app-error.effects';
 import { SnackBarEffects } from './snack-bar.effects';
 import { TaskEffects } from './task.effects';
 import { DialogEffects } from './dialog.effects';
+import { CommentEffects } from './comments.effects';
+import { FilesEffects } from './files.effects';
 
 
-export const effects = [ ProductEffects, AppErrorsEffects, SnackBarEffects, TaskEffects ];
+export const effects = [
+	ProductEffects,
+	AppErrorsEffects,
+	SnackBarEffects,
+	TaskEffects,
+	CommentEffects,
+	FilesEffects,
+];
